pull out remaining usages of case/related
