+++ corehq/ex-submodules/casexml/apps/case/util.py
@@ -1,10 +1,11 @@
 from __future__ import absolute_import
 
 from xml.etree import ElementTree
+
 from django.conf import settings
+
 from casexml.apps.case import const
 from casexml.apps.case.dbaccessors import get_indexed_case_ids
-from casexml.apps.case.sharedmodels import CommCareCaseIndex
 from casexml.apps.phone.models import SyncLogAssertionError, get_properly_wrapped_sync_log
 from casexml.apps.stock.models import StockReport
 from couchforms.models import XFormInstance
@@ -107,18 +108,6 @@ def update_sync_log_with_checks(sync_log, xform, cases, case_db,
                                         case_id_blacklist=case_id_blacklist)
 
 
-def reverse_indices(db, case, wrap=True):
-    kwargs = {
-        'wrapper': lambda r: CommCareCaseIndex.wrap(r['value']) if wrap else r['value']
-    }
-    return db.view(
-        "case/related",
-        key=[case['domain'], case['_id'], "reverse_index"],
-        reduce=False,
-        **kwargs
-    ).all()
-
-
 def get_indexed_cases(domain, case_ids):
     """
     Given a base list of cases, gets all wrapped cases that they reference
@@ -127,21 +116,6 @@ def get_indexed_cases(domain, case_ids):
     from casexml.apps.case.models import CommCareCase
     return [CommCareCase.wrap(doc) for doc in iter_docs(CommCareCase.get_db(),
                                                         get_indexed_case_ids(domain, case_ids))]
-
-
-def get_reverse_indexed_cases(domain, case_ids):
-    """
-    Given a base list of cases, gets all wrapped cases that directly
-    reference them (child cases).
-    """
-    from casexml.apps.case.models import CommCareCase
-    keys = [[domain, id, 'reverse_index'] for id in case_ids]
-    return CommCareCase.view(
-        'case/related',
-        keys=keys,
-        reduce=False,
-        include_docs=True,
-    )
 
 
 def primary_actions(case):
