+++ corehq/ex-submodules/casexml/apps/case/xform.py
@@ -7,11 +7,12 @@ from couchdbkit import ResourceNotFound
 import datetime
 from django.db.models import Q
 import redis
+from casexml.apps.case.dbaccessors import get_reverse_indexed_cases
 from casexml.apps.case.signals import cases_received, case_post_save
 from casexml.apps.phone.cleanliness import should_track_cleanliness, should_create_flags_on_submission
 from casexml.apps.phone.models import OwnershipCleanlinessFlag
 from corehq.toggles import LOOSE_SYNC_TOKEN_VALIDATION
-from casexml.apps.case.util import iter_cases, get_reverse_indexed_cases
+from casexml.apps.case.util import iter_cases
 from couchforms.models import XFormInstance
 from casexml.apps.case.exceptions import (
     IllegalCaseId,
